+++ qutip/cy/br_codegen.py
@@ -27,11 +27,10 @@ class BR_Codegen(object):
                 omp_thresh=None,
                 omp_threads=None,
                 atol=None):
-        try:
-            import cython
-        except (ImportError, ModuleNotFoundError):
+        if not qset.has_cython:
             raise ModuleNotFoundError("Cython is needed for "
                                       "time-depdendent brmesolve")
+        import cython
         import sys
         import os
         sys.path.append(os.getcwd())