+++ qutip/qobj.py
@@ -28,6 +28,7 @@ from qutip import __version__
 from qutip.fastsparse import fast_csr_matrix, fast_identity
 from qutip.cy.ptrace import _ptrace
 from qutip.permute import _permute
+from qutip.dimensions import flatten
 from qutip.sparse import (
     sp_eigs, sp_expm, sp_fro_norm, sp_max_norm, sp_one_norm, sp_L2_norm,
 )
@@ -332,10 +333,15 @@ class Qobj(object):
         # use cases such as enr_fock and other enr_ functions.
         # This does leave open the possibility of data still being
         # misused such as Qobj(complex[n**2][1], dims = [[n],[n]])
-        if (self._data.shape[0] > np.prod(np.hstack(self.dims[0])) or
-           self._data.shape[1] > np.prod(np.hstack(self.dims[1]))) and \
-           self.type != 'super':
-
+        dims_size = (
+            np.prod(flatten(self.dims[0]), dtype=np.int64),
+            np.prod(flatten(self.dims[1]), dtype=np.int64)
+        )
+        shape = self._data.shape
+        if (
+            (shape[0] > dims_size[0] or shape[1] > dims_size[1])
+            and self.type != 'super'
+        ):
             raise ValueError(f"Qobj has smaller dims {self.dims} " +
                              f"than underlying shape {self._data.shape}")
 
@@ -386,8 +392,7 @@ class Qobj(object):
                 out.data = self.data + dat * fast_identity(
                     self.shape[0])
             else:
-                out.data = self.data
-                out.data.data = out.data.data + dat
+                raise TypeError("Only operators can be added to a scalar.")
 
             out.dims = self.dims
 
@@ -415,8 +420,7 @@ class Qobj(object):
             if other.type in ['oper', 'super']:
                 out.data = dat * fast_identity(other.shape[0]) + other.data
             else:
-                out.data = other.data
-                out.data.data = out.data.data + dat
+                raise TypeError("Only operators can be added to a scalar.")
             out.dims = other.dims
 
             if settings.auto_tidyup:
@@ -1357,9 +1361,12 @@ class Qobj(object):
         if sparse:
             q = Qobj()
             q.data, q.dims, _ = _ptrace(self, sel)
-            return q.tidyup() if settings.auto_tidyup else q
+            out = q.tidyup() if settings.auto_tidyup else q
         else:
-            return _ptrace_dense(self, sel)
+            out = _ptrace_dense(self, sel)
+        if isket(out):
+            out = out.proj()
+        return out
 
     def permute(self, order):
         """Permutes a composite quantum object.
@@ -1842,7 +1849,7 @@ class Qobj(object):
 
         """
         if self.isoper:
-            q = Qobj(self.data[states_inds, :][:, states_inds])
+            q = Qobj(self.data[states_inds, :][:, states_inds], isherm=self._isherm or None)
         elif self.isket:
             q = Qobj(self.data[states_inds, :])
         elif self.isbra: