+++ qutip/tests/test_mesolve.py
@@ -1,7 +1,7 @@
 from functools import partial
 
 import numpy as np
-from numpy.testing import assert_, run_module_suite, assert_allclose
+from numpy.testing import assert_, assert_allclose
 import pytest
 
 # disable the MC progress bar
@@ -955,6 +955,22 @@ def test_tlist_h_with_constant_c_ops():
     result = mesolve(hamiltonian, state, few_times, c_ops=[collapse])
     assert result.num_collapse == 1
     assert len(result.states) == len(few_times)
+
+
+def test_mixed_e_ops():
+    """
+    Test callable and Qobj e_ops can mix.
+
+    See gh-2118.
+    """
+    state = basis(2, 0)
+    hamiltonian = QobjEvo(sigmax())
+    collapse = create(2)
+    e_ops = [qeye(2), lambda t, qobj: qobj.norm()]
+    result = mesolve(
+        hamiltonian, state, [0, 1, 2], c_ops=[collapse], e_ops=e_ops
+    )
+    assert result.num_expect == 2
 
 
 def test_tlist_h_with_other_tlist_c_ops_raises():